:lipstick: use withdrawal details component on edit profile page
