+++ app/edit-profile/edit-profile.tsx
@@ -15,10 +15,15 @@ import { TextEditor, useTextEditor } from '@/components/editor'
 import clsx from 'clsx'
 import { Row } from '@/components/layout/row'
 import { Col } from '@/components/layout/col'
-import { redirect } from 'next/navigation'
+import Stripe from 'stripe'
+import { WithdrawalDetails } from '@/components/withdrawal-details'
 
-export function EditProfileForm(props: { profile: Profile }) {
-  const { profile } = props
+export function EditProfileForm(props: {
+  profile: Profile
+  stripeAccount: Stripe.Account | null
+  stripeLoginLink: Stripe.LoginLink | null
+}) {
+  const { profile, stripeAccount, stripeLoginLink } = props
   const { supabase, session } = useSupabase()
   const [username, setUsername] = useState<string>(profile.username)
   const [bio, setBio] = useState<string>(profile.bio)
@@ -258,23 +263,11 @@ export function EditProfileForm(props: { profile: Profile }) {
           setAvatar(event.target.files ? event.target.files[0] : null)
         }}
       ></input>
-      <Button
-        onClick={async () => {
-          const response = await fetch('/api/create-connect-account', {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({
-              profileId: profile.id,
-            }),
-          })
-          const json = await response.json()
-          router.push(json.url)
-        }}
-      >
-        Set up withdrawals
-      </Button>
+      <WithdrawalDetails
+        userId={profile.id}
+        account={stripeAccount}
+        loginLink={stripeLoginLink}
+      />
       <p className="text-center text-rose-500">{errorMessage}</p>
       <Button
         type="submit"
